Clean up the fourier analysis page
